Darken the grays by one shade for Chalk.
Unfortunately, the grays are washed out on Chalk. This results in a very light watchface. So, for Chalk we darken the minute hand from GColorDarkGray to GColorBlack, we darken the 12,3,6,9 dots from Dark Gray to Black, and we darken the clock background from Light Gray to Dark Gray. We *might* be able to use Light Gray for some shadows on the hour plate.